requirements.txt
numpy==2.0.0
pygame==2.6.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
